% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalKmeans.R
\name{balKmeans}
\alias{balKmeans}
\title{Balanced Kmeans function}
\usage{
balKmeans(data, k = 2, standardize = FALSE, max_iter = 20)
}
\arguments{
\item{data}{dataset containing numerical columns of observations to be clustered.}

\item{k}{number of output clusters required; default=2.}

\item{standardize}{standardization of observation values required; by default=FALSE.}

\item{max_iter}{maximum iterations to be made for clustering; default=20.}
}
\value{
A dataframe object with Balanced k clusters
}
\description{
This function implements balanced K-Means algorithm where the minimum and maximum number of observations in each cluster and the power ratings of each observation are taken as constraints.
}
