% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read CSV Sample}
\usage{
sample_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
a \code{tibble}
}
\description{
Read CSV Sample
}
\examples{
csv = system.file("extdata","school_coordinates.csv",package="BalKmeans")
sample_read(csv)
}
